+++ src/components/App/App.js
@@ -1,0 +1,9 @@
+import './App.css';
+
+function App() {
+  return (
+    null
+  );
+}
+
+export default App;
